+++ Classes/Controller/UserWithoutAutologinController.php
@@ -13,8 +13,10 @@ use OliverKlee\Onetimeaccount\Service\CaptchaFactory;
 use OliverKlee\Onetimeaccount\Service\CredentialsGenerator;
 use OliverKlee\Onetimeaccount\Validation\CaptchaValidator;
 use OliverKlee\Onetimeaccount\Validation\UserValidator;
+use Psr\Http\Message\ResponseInterface;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Annotation\IgnoreValidation;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
 use TYPO3\CMS\Extbase\Validation\Validator\ConjunctionValidator;
 use TYPO3\CMS\Frontend\Authentication\FrontendUserAuthentication;
@@ -56,9 +58,9 @@ class UserWithoutAutologinController extends ActionController
     /**
      * Creates the user creation form (which initially is empty).
      *
-     * @TYPO3\CMS\Extbase\Annotation\IgnoreValidation("user")
+     * @IgnoreValidation("user")
      */
-    public function newAction(?FrontendUser $user = null, ?int $userGroup = null): void
+    public function newAction(?FrontendUser $user = null, ?int $userGroup = null): ResponseInterface
     {
         $newUser = ($user instanceof FrontendUser) ? $user : GeneralUtility::makeInstance(FrontendUser::class);
         $this->view->assign('user', $newUser);
@@ -81,6 +83,7 @@ class UserWithoutAutologinController extends ActionController
         if (\is_string($redirectUrl) && $redirectUrl !== '') {
             $this->view->assign('redirectUrl', $redirectUrl);
         }
+        return $this->htmlResponse();
     }
 
     public function initializeCreateAction(): void
@@ -115,8 +118,11 @@ class UserWithoutAutologinController extends ActionController
      *
      * @throws \RuntimeException
      */
-    public function createAction(?FrontendUser $user = null, ?int $userGroup = null, ?Captcha $captcha = null): void
-    {
+    public function createAction(
+        ?FrontendUser $user = null,
+        ?int $userGroup = null,
+        ?Captcha $captcha = null
+    ): ResponseInterface {
         if (!$user instanceof FrontendUser) {
             return;
         }